Stop using the d2to1-based pbr
pbr updated itself to remove the d2to1 dependency. One of the main
reasons for this is a direct upstream dependency on distribute, which
causes the distribute/setuptools merge bug to get triggered, which
breaks people's system if they install with pip < 1.4.

Remove d2to1 references, and also bump the required version requested of
pbr to a version late enough to no longer depend on d2to1.

Remove the version specifier from setup.py, because setup_requires has
no way to trigger a version requirement upgrade. The flat version should
be sufficient for the setup.py interface in perpetuity.

Remove the setup-hook from setup.cfg, as it is no longer needed.

Change-Id: I40d3507c77802d13fd894aa16714b2ed6aa7d59f