uncap SQLA in requirements.txt
migrate can't just take a global requirements sync because it
needs to be tested against multiple versions of SQLA to assure
compatibility. A recent change had the effect of only testing
migrate against SQLA 0.7, which is definitely *not* what we
want to be doing.

this reverts that change, and leaves very specific comments to
hopefully prevent this from happening in the future.

Change-Id: Icb4e136f0de6caa224019bb955341c4b67c5e1a1