Use a modern PBR package
The 2.0.0 is breaking in that it removes the use of warnerrors in
build_sphinux.

sqlalchemy-migrate isn't using that feature, so it shoudln't break.

The cap on pbr is preventing other OpenStack projects that would like to
use pbr 2.0.0 (and sphinx 1.5.1) from doing so as it breaks
co-installability with sqlalchemy-migrate

Change-Id: I1c907201c717fe42caca24831985a119f2a1738b
Related-Bug: 1668848